lang: qdecode_t, the expansion half of the qdot decode family
Same decode as <fmt>_qdot_t, stored to the [K, N] scratch a batched matmul
reads instead of contracted. Shared with the contraction rather than written
twice; the thread map is what differs, and the store sets it.

The tables are packed to one i8 a slot so a lane's entry is one vector load,
and the matvecs read their activations as vector<4xf32>. Both cut a lane's
memory ops per k-block from 20 to 7.

+++ phobos-lang/src/codegen/tests/iq3s_qdot.rs
@@ -7,8 +7,8 @@ const SRC: &str = "\
 @autotune(TN in [8])
 @aligned(N = TN)
 kernel iq3s_qdot_matvec(A: tensor<f32>[M, K], QB: tensor<i8>[N, RB],
-                        D: tensor<f16>[N, NB], GRID: tensor<i32>[1, 2048],
-                        SIGNS: tensor<i32>[1, 2048], C: tensor<f32>[M, N]) {
+                        D: tensor<f16>[N, NB], GRID: tensor<i8>[1, 2048],
+                        SIGNS: tensor<i8>[1, 2048], C: tensor<f32>[M, N]) {
   let pn = program_id(0)
   C[0 :+ 1, pn * TN :+ TN] = iq3s_qdot_t(A[0 :+ 1, :], QB[pn * TN :+ TN, :],
                                          D[pn * TN :+ TN, :], GRID[0 :+ 1, :], SIGNS[0 :+ 1, :])
@@ -46,8 +46,8 @@ fn iq3s_qdot_t_reads_a_dynamic_contraction_length_at_runtime() {
 fn iq3s_qdot_t_rejects_a_ragged_contraction() {
     let src = "@launch(256)
         kernel iq3s_decode(A: tensor<f32>[1, 128], QB: tensor<i8>[8, 55],
-                           D: tensor<f16>[8, 1], GRID: tensor<i32>[1, 2048],
-                           SIGNS: tensor<i32>[1, 2048], C: tensor<f32>[1, 8]) {
+                           D: tensor<f16>[8, 1], GRID: tensor<i8>[1, 2048],
+                           SIGNS: tensor<i8>[1, 2048], C: tensor<f32>[1, 8]) {
             C[0 :+ 1, 0 :+ 8] = iq3s_qdot_t(A[0 :+ 1, :], QB[0 :+ 8, :], D[0 :+ 8, :], GRID[0 :+ 1, :], SIGNS[0 :+ 1, :])
         }";
     let err = std::panic::catch_unwind(|| emit_mlir(src));
